Refactor: Move validation methods to PaymentAccountValidation
